renaming first test suite
